% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_number_valid.R
\name{is_number_valid}
\alias{is_number_valid}
\title{Checks validity of an NHS number}
\usage{
is_number_valid(nhs_number, warn = TRUE)
}
\arguments{
\item{nhs_number}{A 10 digit NHS number to validate}

\item{warn}{Boolean that controls display of warning messages}
}
\description{
NHS numbers are 10 digit numbers where the 10th digit is a checksum.
This function validates the checksums of the supplied NHS numbers.
}
\examples{
is_valid(1234567881)
is_valid(1234567890, warn = FALSE)
}
